alignment of arguments types for std::min

+++ c++/src/Int128.cc
@@ -22,6 +22,7 @@
 #include <iomanip>
 #include <iostream>
 #include <sstream>
+#include <algorithm> // std::min
 
 namespace orc {
 
@@ -41,7 +42,7 @@ namespace orc {
       bool isNegative = str[0] == '-';
       size_t posn = isNegative ? 1 : 0;
       while (posn < length) {
-        size_t group = std::min(18ul, length - posn);
+        size_t group = std::min(static_cast<size_t>(18), length - posn);
         int64_t chunk = std::stoll(str.substr(posn, group));
         int64_t multiple = 1;
         for(size_t i=0; i < group; ++i) {
